docs: add engine agent guide
